refactor: Use PEP 8 module names
we had an error that was like:

```python3
import Drive
from Drive import Drive
```

For reasons like this, it's super important to follow the spec for langs that prefer runtime checks...
