+++ examples/wheels.py
@@ -6,7 +6,7 @@ os.chdir(os.path.dirname(os.path.abspath(__file__)))
 import sys
 sys.path.append('../')
 
-from libs import UDPOut
+from libs import udp_out
 
 HOST = '192.168.1.101' 
 # 127.0.0.1 is the 'loopback' address, or the address
@@ -16,7 +16,7 @@ PORT = 1001
 # this number is arbitrary as long as it is above 1024
 speed = 90
 
-UDPOut.sendWheelSpeeds(HOST, PORT, speed, speed, speed, speed, speed, speed)
+udp_out.send_wheel_speeds(HOST, PORT, speed, speed, speed, speed, speed, speed)
 # the six arguments represent each wheels speed (on a scale between 0 and 255, since it will be stored in a byte).
 # the order of the wheels in the arguments is Front Left, Middle Left, Rear Left, Front Right, Middle Right, and Rear Right. 
 # sendWheelSpeeds(HOST, PORT, fl, ml, rl, rt, mr, rr)
@@ -25,5 +25,5 @@ if __name__ == "__main__":
     print("Sup")
     speed = 90
     while True:
-        UDPOut.sendWheelSpeeds(HOST, PORT, speed, speed, speed, speed, speed, speed)
+        udp_out.send_wheel_speeds(HOST, PORT, speed, speed, speed, speed, speed, speed)
         sleep(1)
