+++ libs/aruco_tracker.py
